commit_53173 ------------------------------------------------------------------------ r53173 | vboxsync | 2014-11-02 07:29:05 +0800 (Sun, 02 Nov 2014) | 1 line
IPRT,SUPDrv: Added new 'ringbuf[=size]' log destination for cases where only the last bit of the log is interesting and we can save space and time by not writing the whole log to disk/wherever.  Since it changes the internal logger structure, it'll require a major SUPDrv version bump.
------------------------------------------------------------------------