commit_53178 ------------------------------------------------------------------------ r53178 | vboxsync | 2014-11-03 05:04:12 +0800 (Mon, 03 Nov 2014) | 1 line
HMVMXR0.cpp: Fixed several bugs in the real-mode-on-top-of-v86 instruction emulation: STI always set the inhibit interrupts. IRET used didn't keep enough flags in the high word, but incorrectly (I believe) kept the resume flags (RF). The RF flag was only cleared in by a few instructions, it should normally be cleared when an instruction is retired (with a couple of exception). Made the real-mode direct event injection code return VINF_EM_DBG_STEPPED when stepping (currently passing fStepping around as that clearer).
------------------------------------------------------------------------