commit_53308 ------------------------------------------------------------------------ r53308 | vboxsync | 2014-11-12 11:00:16 +0800 (Wed, 12 Nov 2014) | 6 lines
NAT: m_adj() - when truncating by exactly the length of the last mbuf
in a chain, don't leave that last mbuf around with m_len set to 0.
Fragmentation code in ip_output0() sending a chain of two mbufs
triggers this for the first fragment and if_encap() drops it thinking
it was passed a chain (or triggers an assertion in a debug build).

------------------------------------------------------------------------