commit_53379 ------------------------------------------------------------------------ r53379 | vboxsync | 2014-11-24 07:19:31 +0800 (Mon, 24 Nov 2014) | 1 line
Config.kmk: Two fixes that avoids using the wrong compilers when VBOX_WITH_MACOSX_COMPILERS_FROM_DEVEL (or tools/darwin.x86/xcode/v*) is used. The TOOL_GXX32 is for raw-mode on 64-bit host. The dtrace hack is for preprocessing (defaults to clang, no way of giving the replacement a -issyroot, but this seems to work for Xcode 4.1 tools).
------------------------------------------------------------------------