commit_53775 ------------------------------------------------------------------------ r53775 | vboxsync | 2015-01-12 23:54:53 +0800 (Mon, 12 Jan 2015) | 1 line
DevVGA-SVGA.cpp: Fixed bugs in SVGA_REG_BITS_PER_PIXEL, SVGA_REG_HEIGHT, and SVGA_REG_HEIGHT write handlers that was uncovered by enabling the code in ring-0 and raw-mode contexts. The device state was changed before we decided that we needed to handle the request in ring-3, and because of the state change nothing was actually done in ring-3.
------------------------------------------------------------------------