commit_54027 ------------------------------------------------------------------------ r54027 | vboxsync | 2015-01-29 23:23:36 +0800 (Thu, 29 Jan 2015) | 2 lines
Main/cbinding: Add new helper functions to clear out strings (useful if they contain sensitive data). While doing this, added code to the glue init which checks if the version of the structure is compatible with what the client expects. This exposed annoying code duplication between the non-Windows/Windows implementations which is now eliminated.

------------------------------------------------------------------------