commit_54116 ------------------------------------------------------------------------ r54116 | vboxsync | 2015-02-09 23:29:23 +0800 (Mon, 09 Feb 2015) | 7 lines
NAT: icmpwin_ping - don't decrement TTL once again, this is done in
ip_input(), where time exceeded is also handled.

Note, that this is not correct, since time exceeded should not be
generated for packets with TTL 1 that are for "us", i.e. one of the
CTL_ALIAS addresses.  But for now just do as the Romans do.

------------------------------------------------------------------------