commit_54224 ------------------------------------------------------------------------ r54224 | vboxsync | 2015-02-17 06:41:32 +0800 (Tue, 17 Feb 2015) | 1 line
SUP,IPRT: Started as a build fix ended up as a cleanup (mostly untested as I'm windows). We'll keep SUP_IOCTL_TSC_READ for the purpose of dealing with the i64TSCDelta == INT64_MAX in SUPReadTsc/SUPReadTscWithDelta.  Moved the excessive inline code from sup.h to SUPLibAll.cpp (new file) and SUPDrv.c (apply delta).  Don't bother too much about trying to share code between SUPLibAll.cpp and SUPDrv.c/++ as the SUPReadTsc[WithDelta] scenario does not care about knowing when things goes bad, it just needs a TSC that is as good as we can get.  The SUPDrv.c code on the other hand probably needs other kind of status codes, assertions and whatnot to be better off with its own version of the code.  Remove the now unused assembly macro for applying the delta - the way the delta is applied will not change, period, so better just document it instead.
------------------------------------------------------------------------