commit_54339 ------------------------------------------------------------------------ r54339 | vboxsync | 2015-02-21 02:10:12 +0800 (Sat, 21 Feb 2015) | 1 line
SUPDrv,VMM: Added SUPR0TscDeltaMeasureBySetIndex to SUPDrv (bumping version req) so that VMMR0.cpp can, if necessary, trigger a delta measurement prior to doing RC and HM stuff. Also, made the context hook set the return-to-ring3 force flag if we're rescheduled on a CPU with an TSC delta needing measuring.  This is probably code that will almost never get called...
------------------------------------------------------------------------