commit_54799 ------------------------------------------------------------------------ r54799 | vboxsync | 2015-03-17 05:23:03 +0800 (Tue, 17 Mar 2015) | 1 line
CPUM: Fixed two problems with the CPUID code. cLeaves wasn't updated in cpumR3CpuIdInstallAndExplodeLeaves leading to garbage entries at the end of the array. And when the CPUID insertion code had a bug when inserting subleafs that triggered when doing the cache leaf (4) fake up during old saved state loading.
------------------------------------------------------------------------