commit_55967 ------------------------------------------------------------------------ r55967 | vboxsync | 2015-05-20 21:12:28 +0800 (Wed, 20 May 2015) | 2 lines
Main/Host: Close the support driver as soon as we no longer need it. Not worth keeping it open all the time. Allows reloading the driver while VBoxSVC is running, and avoids quirks during debugging of VBoxSVC: it is quite normal that the Host object is created several times, which means that multiple SUPR3InitEx calls are made.

------------------------------------------------------------------------