commit_56529 ------------------------------------------------------------------------ r56529 | vboxsync | 2015-06-18 23:15:17 +0800 (Thu, 18 Jun 2015) | 1 line
tstVDIo: Change the implementation to calculate the throughput in KB/s. Before it would return ~(uint64_t)0 on at least one testbox causing Out of Range errors with the SQL database when inserting that value and returning stale data in the web interface. The current approach while still not very sophisticated should avoid that and the error margin introduced is far below the normal fluctuations during the tests
------------------------------------------------------------------------