commit_56532 ------------------------------------------------------------------------ r56532 | vboxsync | 2015-06-19 01:44:54 +0800 (Fri, 19 Jun 2015) | 2 lines
doc/manual: fix SDK reference to be valid docbook XML, needed adjustments in the generator (which produced creative stuff which was severely violating the DTD, e.g. custom attributes which could be easily done properly) and a LOT of manual editing. Content of the manual is essentially the same, the biggest issues were <para> sloppiness and the use of <xref> where <link> is the only solution. Also made the line lengths again as uniform as possible.

------------------------------------------------------------------------