commit_56658 ------------------------------------------------------------------------ r56658 | vboxsync | 2015-06-26 20:51:57 +0800 (Fri, 26 Jun 2015) | 1 line
Storage/AHCI: Fix possible use after free after an I/O error occurred. The first I/O request with an I/O error is saved for the error log page but is also put on the per port free requerst list. When the guest reads the error page the request is freed but is still on the free request list and could be used later
------------------------------------------------------------------------