commit_57054 ------------------------------------------------------------------------ r57054 | vboxsync | 2015-07-22 19:38:49 +0800 (Wed, 22 Jul 2015) | 2 lines
Devices/Audio/DrvAudio.cpp: Fix cleanup logic, as having it in the destruct callback is too late (it uses the driver below, which is already destructed at this point). Kind of a use after free, but for VM poweroff the memory is freed later (it crashed when constructing failed as then the driver below is freed before destructing this one).

------------------------------------------------------------------------