commit_57135 ------------------------------------------------------------------------ r57135 | vboxsync | 2015-07-31 03:41:09 +0800 (Fri, 31 Jul 2015) | 1 line
vmsvga3d/darwin/ogl: Resize the shared context, save and restore the shared context. Don't try too hard to do the NSOpenGLContext::setView thing if it doesn't want to work, it's pretty noisy on 10.8.5/nvidia. Do the vmsvga3dPowerOn on EMT, not the FIFO when restoring state (darwin only atm).
------------------------------------------------------------------------