commit_57267 ------------------------------------------------------------------------ r57267 | vboxsync | 2015-08-11 09:27:26 +0800 (Tue, 11 Aug 2015) | 8 lines
dbgcTcpBackRead: EOF on socket is reported as success by RTTcpRead()
if pcbRead is non-NULL, so check if we successfully read 0 bytes and
return VERR_NET_SHUTDOWN in that case (like RTTcpRead() already does
when pcbRead is NULL).

Makes it possible to disconnect from the debugger without causing a
busy loop in dbgcInputRead().

------------------------------------------------------------------------