commit_57303 ------------------------------------------------------------------------ r57303 | vboxsync | 2015-08-13 08:31:23 +0800 (Thu, 13 Aug 2015) | 1 line
SUPDrv-win.cpp: Workaround for AudioDG.exe on windows 10 wanting to do some new cpu set management for us, not sure what it is or what I think of it.  We need audio though, so allow it for now.  Also modified the SUPDRVNTPROTECT handling to make sure we don't try remove structures not in the tree, in case we fail early before it's inserted.  Finally, shut up a signed comparisons warning.
------------------------------------------------------------------------