commit_57456 ------------------------------------------------------------------------ r57456 | vboxsync | 2015-08-19 22:37:00 +0800 (Wed, 19 Aug 2015) | 1 line
SUPDrv-win.cpp: Windows 10 VERR_LDR_MISMATCH_NATIVE fix. Turns out the ImageBase member of the optional header is sometimes updated to the actual load address by the NT kernel or some associate. Why this happens is unclear, but it's easy to deal with.  The workaround allows for both the kernel and iprt to update the ImageBase on load, just in case we change IPRT some time in the future.
------------------------------------------------------------------------