commit_57790 ------------------------------------------------------------------------ r57790 | vboxsync | 2015-09-17 01:09:28 +0800 (Thu, 17 Sep 2015) | 1 line
VMSVGA3d: Turns out I'm blitting to the wrong end (Y direction) of the nsview framebuffer. Missed some [NSOpenGLContext update] calls when the parent view got resized, so it sometimes worked, but not all the time. After calling update, the visible part of the framebuffer is the bottom half, not the top half as I though and as seems to be the case on the other plaforms.
------------------------------------------------------------------------