commit_57882 ------------------------------------------------------------------------ r57882 | vboxsync | 2015-09-24 22:40:39 +0800 (Thu, 24 Sep 2015) | 1 line
Serial/NamedPipe: Reorder error checks in the receive code path on Windows. Looks like Kaspersky AV changes the behavior of the named pipe API when hooking into it, always returning ERROR_IO_PENDING on an overlapped read even if no other process connected to the named pipe yet. ERROR_PIPE_LISTENING is returned as the overlapped result making our error checks trip and aborting the receive thread. The fix moves the check for ERROR_IO_PENDING to the front so it executed first and the check for other non fatal errors is done afterwards
------------------------------------------------------------------------