commit_57906 ------------------------------------------------------------------------ r57906 | vboxsync | 2015-09-26 04:43:36 +0800 (Sat, 26 Sep 2015) | 1 line
RTProc/win,RTErr: Avoid doing LogonUserW when we don't need to get the profile environment as this may require the TCB privilege. On W2K CreateProcessWithLogonW doesn't plant standard handles in the child process, so try our best to do it ourselves. (This may be applicable to later windows versions too, but behavior must've been change in W10 or earlier.) Introduced specific error codes for the missing process rights to prevent wasting time.
------------------------------------------------------------------------