commit_57979 ------------------------------------------------------------------------ r57979 | vboxsync | 2015-10-01 16:25:21 +0800 (Thu, 01 Oct 2015) | 1 line
Parallel ports: Several fixes for multiple parallel ports. Make the second parallel port visible to the guest through ACPI if enabled (and don't expose the first port if it is not enabled), increase the maximum instanbce count to 2 and make it possible to enable a parallel port without having it connected to a host device to make it behave like it is not connected to anything.
------------------------------------------------------------------------