commit_58619 ------------------------------------------------------------------------ r58619 | vboxsync | 2015-11-09 23:13:19 +0800 (Mon, 09 Nov 2015) | 2 lines
bugref:8087: Additions/x11: support non-root X server: make the VBoxOGL.so replacement OpenGL library executable. Executing it returns 0 if the system it is running on supports VirtualBox OpenGL pass-through in its current configuration and 1 otherwise. The intended use is knowing at guest boot time whether or not to put the library in place. Makefile parts.

------------------------------------------------------------------------