commit_58622 ------------------------------------------------------------------------ r58622 | vboxsync | 2015-11-10 03:02:19 +0800 (Tue, 10 Nov 2015) | 1 line
bugref: 8087:Additions/x11: support non-root X server: make the VBoxOGL.so replacement OpenGL library executable.  Executing it returns 0 if the system it is running on supports VirtualBox OpenGL pass-through in its current configuration and 1 otherwise.  The intended use is knowing at guest boot time whether or not to put the library in place.  Makefile parts.  Fix: this only applies to Linux guests.
------------------------------------------------------------------------