commit_58628 ------------------------------------------------------------------------ r58628 | vboxsync | 2015-11-10 09:25:13 +0800 (Tue, 10 Nov 2015) | 1 line
bs3kit: Managed to link 16, 32 and 64-bit C code into the same binary, where the 64-bit C code was produced by the microsoft compiler.  That doesn't mean it actually works, but chances are high.  Bad new though, no debug info.  OTOH, a lot of the assembly code can be converted to C, which is faster and safer than adjusting it to the new calling convensions.
------------------------------------------------------------------------