commit_58660 ------------------------------------------------------------------------ r58660 | vboxsync | 2015-11-11 20:36:35 +0800 (Wed, 11 Nov 2015) | 2 lines
bugref:8087: Additions/x11: support non-root X server: when we build our OpenGL library the standard way (not hijacking Mesa) we have problems with SELinux on Fedora due to our assembly stubs which require text relocations. We have code to do this with C or assembly, and I do not know any reason why assembly is better, so switch to C until I do.

------------------------------------------------------------------------