commit_58672 ------------------------------------------------------------------------ r58672 | vboxsync | 2015-11-12 18:03:51 +0800 (Thu, 12 Nov 2015) | 1 line
bugref:8051: Installers: unify Linux host installers as far as possible: sometimes running vboxdrv.sh setup as called by loadall.sh caused the modules to be built twice.  This was because certain work was duplicated between the call to stop && setup && start, and the implementation of start().  This change removes the duplication.
------------------------------------------------------------------------