commit_58891 ------------------------------------------------------------------------ r58891 | vboxsync | 2015-11-27 08:28:57 +0800 (Fri, 27 Nov 2015) | 1 line
dbgf.h: Extended DBGFEVENTTYPE with a bunch of events that may be useful.  The CPU exceptions (DBGFEVENT_XCPT_XXX) are of immediate use. Events starting with DBGFEVENT_FIRST_SELECTABLE are envisioned being configured via a single API operating on some bitmask and making HM, IOM, ++ notifications calls to let the components update cached state/whatever.  The interrupt events would require it's own API, that operates on a 32 byte bitmap with a bit counter for optimizations.  The I/O and MMIO events should be added to the DBGFR3BpXxx API and added as two new DBGFBP variants.
------------------------------------------------------------------------