commit_58897 ------------------------------------------------------------------------ r58897 | vboxsync | 2015-11-27 18:41:22 +0800 (Fri, 27 Nov 2015) | 1 line
bugref:8136: OL7.2: VBoxService not started: wake up systemd when we create or recreate a systemd unit so that systemd can refresh its SELinux policy information.  This works around a problem in systemd whereby it refuses to enable or start an updated unit, though the change is probably the right thing to do in any case.
------------------------------------------------------------------------