commit_58993 ------------------------------------------------------------------------ r58993 | vboxsync | 2015-12-05 00:26:50 +0800 (Sat, 05 Dec 2015) | 1 line
bugref:8153: vboxvideo: ABI issues / software rendering on Linux guests: add a new service to VBoxClient to test whether accelerated 3D pass-through is currently available by trying to connect to the HGCM service and signalling success or failure in the process exit code.  Add generic HGCM connect and disconnect functionality to VBoxGuestR3Lib to support this.  Add the concept of non-daemon services to VBoxClient as we do not need to daemonise or create a pid-file for this quick test.
------------------------------------------------------------------------