commit_59117 ------------------------------------------------------------------------ r59117 | vboxsync | 2015-12-14 22:04:37 +0800 (Mon, 14 Dec 2015) | 1 line
USB,Main: Rework USBProxyService. Split it into a USBProxyService and USBProxyBackend class, USBProxyService can use multiple USBProxyBackend instances as sources for USB devices to attach to a VM which will be used for USB/IP support. Change the PDM USB API to contain a backend parameter instead of a remote flag to indicate the USB backend to use for the given device.
------------------------------------------------------------------------