commit_59166 ------------------------------------------------------------------------ r59166 | vboxsync | 2015-12-17 03:53:20 +0800 (Thu, 17 Dec 2015) | 1 line
bugref:8087: Additions/x11: support non-root X server: on Linux, correctly set the shared object name of our library to libGL.so.1, to make sure that the dynamic linker cacher handles it correctly.  This had slipped into the make the VBoxOGL.so replacement OpenGL library executable changes which were reverted, but it should have been a separate change from the start - this is now fixed.
------------------------------------------------------------------------