commit_59234 ------------------------------------------------------------------------ r59234 | vboxsync | 2015-12-31 18:41:47 +0800 (Thu, 31 Dec 2015) | 1 line
bugref:8087: Additions/x11: support non-root X server: not really necessary for this task, but do set-up and clean-up of the X11 and graphics parts of the Additions from the main Additions service script.  This also allows us to share a system image more easily between virtualised and non-virtulaised systems by preventing graphics set-up if we see that we are not running inside a VirtualBox guest.
------------------------------------------------------------------------