commit_59248 ------------------------------------------------------------------------ r59248 | vboxsync | 2016-01-04 22:13:22 +0800 (Mon, 04 Jan 2016) | 1 line
Storage: Get rid of the block driver and merge the the little extra functionality it had into the VD driver. Enables us to get rid of PDMIBLOCK which is basically a subset of PDMIMEDIA and makes changes to the latter interface tedious because it had to be replicated in the former. (bugref:4114)
------------------------------------------------------------------------