commit_59307 ------------------------------------------------------------------------ r59307 | vboxsync | 2016-01-11 19:37:12 +0800 (Mon, 11 Jan 2016) | 1 line
bugref:8183: Mouse Integration sometimes not working after restore from saved state: fix a mishandling of the guest cursor shape protocol in the graphics device.  When the guest sends a cursor shape event with the visible flag set but no shape data, the value of the alpha flag should be ignored.  We were indeed ignoring it, but still saving its (usually incorrect) value to the device context, overwriting the old value, and from there to the saved state.
------------------------------------------------------------------------