commit_60177 ------------------------------------------------------------------------ r60177 | vboxsync | 2016-03-24 03:12:39 +0800 (Thu, 24 Mar 2016) | 1 line
bugref:8087: Additions/x11: support non-root X server: add a capability which allows guest drivers to tell the host that they do not depend on the settings of the legacy VBE registers.  Windows guest blue screens cause the VBE registers to change, and do not work correctly if the host does not handle this.
------------------------------------------------------------------------