commit_60198 ------------------------------------------------------------------------ r60198 | vboxsync | 2016-03-26 22:43:03 +0800 (Sat, 26 Mar 2016) | 1 line
bugref:8087: Additions/x11: support non-root X server: when running on older versions of VirtualBox certain things are not shown right unless we programme the graphics mode on the first screen via the VBE registers too.  While this is not always possible due to limitations of the register model, at least do it when it is.
------------------------------------------------------------------------