commit_60279 ------------------------------------------------------------------------ r60279 | vboxsync | 2016-04-01 02:57:37 +0800 (Fri, 01 Apr 2016) | 1 line
ValidationKit/usb: Early commit of the new remote USB test configuration and execution server. It will replace the current approach of using TXS to configure the gadgets because that approach is not suitable for multiple clients accessing the same gadget (software devices exported over USB/IP). Non functional and heavily work in progress
------------------------------------------------------------------------