commit_60288 ------------------------------------------------------------------------ r60288 | vboxsync | 2016-04-01 21:14:52 +0800 (Fri, 01 Apr 2016) | 2 lines
Main/Medium+Snapshot: fix long standing bugs in medium locking during merge operations (for backwards merge the target was only locked for reading, confusing the snapshot code) and when canceling a merge (which incorrectly handled media which would have been deleted in the case when a VM is running). The snapshot code had an error handling bug, which fortunately prevented the medium locking issue to result in unsafe updates. Also enabled multiple error reporting in the snapshot deletion code, to get more error context (or not losing as much) if users report issues.

------------------------------------------------------------------------