commit_60317 ------------------------------------------------------------------------ r60317 | vboxsync | 2016-04-05 04:41:07 +0800 (Tue, 05 Apr 2016) | 1 line
bugref:8087: Additions/x11: support non-root X server: in the X.Org video driver, check whether the kernel driver is loaded in PCIProbe(), not in PreInit().  If PCIProbe() fails, the server continues to try the next driver (modesetting in this case), whereas if PreInit() fails it aborts altogether.  This simplifies our life: now we can install the X.Org driver unconditionally and it will just not be used if the kernel driver is there (or the hardware is not), but will be used if it can be and the kernel driver is loaded too late.  This should solve problems when starting Ubuntu 12.04 guests, where we previously tried to load the kernel driver and install the X.Org driver exactly if it failed to load.  We sometimes ended up with both loaded and sometimes with neither.
------------------------------------------------------------------------