commit_60627 ------------------------------------------------------------------------ r60627 | vboxsync | 2016-04-21 21:47:37 +0800 (Thu, 21 Apr 2016) | 2 lines
Main/Medium: When creating a medium lock list, explicitly state which object should be locked for writing (in some case it isn't the object for which the lock list is requested). This is especially important for backward medium merge operation, where the target isn't that object for backward merging. All other changes are simply adjusting code to the changed method parameter list.

------------------------------------------------------------------------