commit_61000 ------------------------------------------------------------------------ r61000 | vboxsync | 2016-05-17 20:08:55 +0800 (Tue, 17 May 2016) | 10 lines
NAT: Initialize guest address guess for the port-forwarding to
CTL_GUEST.  This is strictly an improvement since it helps the common
case and doesn't affect corner cases, which will still need a sign of
life packet just like before.

When a VM is restored we do flap the link for 5 seconds to make guest
reacquire the DHCP lease, but from the guest PoV that flap may be
shorter and sometimes it just ignores it, preventing port-forwarding
from working.  ticketref:15412

------------------------------------------------------------------------