commit_61144 ------------------------------------------------------------------------ r61144 | vboxsync | 2016-05-24 06:16:26 +0800 (Tue, 24 May 2016) | 1 line
CPUM,HM,GVMM,TRPM,VMM: Next part of the FPU state handling for IEM. This is a little bit risky change as we now leave CR0.TS+EM cleared after saving the host state, they only get restored when we restore the host state.  On Windows, Darwin, and later on Linux (needs testing) we will rely on #NM handling of the host OS and not our own CR.TS/EM handy work.  This means we won't be saving the host state but rather the ring-3 state of our own thread.  This change also introduces a CPUM force flag that we're using for restoring CR0.TS/EM in raw-mode (it may be extended with other uses later if we need to).
------------------------------------------------------------------------