commit_61201 ------------------------------------------------------------------------ r61201 | vboxsync | 2016-05-26 09:49:38 +0800 (Thu, 26 May 2016) | 6 lines
NAT: Drop packets to multicast and reserved destinations earlier,
so that we don't spam TTL exceeded messages in response to multicast
membership report.  Only decrement TTL if a packet is not for us.
If a packet is for another node of our private network, drop it.
The latter should address ticketref:15303, though at IP level.

------------------------------------------------------------------------