commit_61597 ------------------------------------------------------------------------ r61597 | vboxsync | 2016-06-09 14:24:00 +0800 (Thu, 09 Jun 2016) | 10 lines
bugref:4567: Linux kernel driver maintenance
drm/vboxvideo: Add delayed update to support fbdev

Due to the virtrual nature of the emulated hardware, the
hardware needs help to know about updates to screen content.
The fb layer provides this with 'deferred IO'.
This patch adds support for this to the vboxvideo DRM driver.

Signed-off-by: Egbert Eich <eich@suse.de>
Signed-off-by: Egbert Eich <eich@suse.com>
------------------------------------------------------------------------