commit_61978 ------------------------------------------------------------------------ r61978 | vboxsync | 2016-07-01 21:42:28 +0800 (Fri, 01 Jul 2016) | 1 line
Storage/iSCSI: Fix a possible crash during VM resume when connectin to the target fails in iscsiSetOpenFlags (readonly -> read/write transition). Failing the connection attempt here leaves the image data in an inconsistent state leading to memory corruption later on if iscsiWrite/iscsiRead is called. Don't do any communication with the target in iscsiSetOpenFlags but use the remembered data from the first login to decide whether the transition is allowed. Introduce a new flag to indicate whether a reconnect attempt should be made from the I/O worker thread before processing new requests to try reconnecting after a VM resume which was suspended due to an unresponsive target
------------------------------------------------------------------------