commit_62014 ------------------------------------------------------------------------ r62014 | vboxsync | 2016-07-05 00:18:43 +0800 (Tue, 05 Jul 2016) | 2 lines
Main/Machine: add a stub for an aux settings file (which will be an additional file, for the current state and each snapshot), meant for stuff which needs to be saved potentially frequently while the VM is running. It should cover what aren't true settings, in the sense of not being subject to the usual saveSettings/discardSettings machinery. Usually rather bulky stuff.

------------------------------------------------------------------------