commit_62170 ------------------------------------------------------------------------ r62170 | vboxsync | 2016-07-12 02:14:11 +0800 (Tue, 12 Jul 2016) | 2 lines
XPCOM: Change the shutdown order in nsExceptionService, attempting to fix a sporadic crash. The biggest change is postponing the lock freeing to the destructor. The order change of the operations in the Shutdown method eliminates the concerns mention in the ThreadDestruct method. The change in DoDropThread is pure paranoia, intended to avoid potential crashes due to continuing with a NULL pointer in *emp if the thread is not known.

------------------------------------------------------------------------