commit_62201 ------------------------------------------------------------------------ r62201 | vboxsync | 2016-07-12 22:14:55 +0800 (Tue, 12 Jul 2016) | 1 line
bugref:8087: Additions/x11: support non-root X server: improve preventing loading of OpenGL libraries by the X server.  We support client-side OpenGL, but not server-side.  Now we intercept functions that the X server will call and check for it rather than trying to overwrite the libraries in memory with the Mesa OpenGL ones.  Also fix locating the Mesa libraries.  Hopefully this can be removed altogether, but not just before the release.
------------------------------------------------------------------------