commit_62319 ------------------------------------------------------------------------ r62319 | vboxsync | 2016-07-19 19:40:50 +0800 (Tue, 19 Jul 2016) | 1 line
Debugger/Digger/Linux: Implement extraction and parsin of the compressed kernel config found in most kernels. The config is guarded by IKCFG_ST and IKCFG_ED for the start and end of the config. The config database will be used later to determine the method to use to find the kernel symbol table.
------------------------------------------------------------------------