commit_62396 ------------------------------------------------------------------------ r62396 | vboxsync | 2016-07-21 21:16:07 +0800 (Thu, 21 Jul 2016) | 1 line
bugref:6947: FE/Qt: Problem with "Alt" key up event in Windows guests when "Alt-Tab"-ing away from guest: switch from injecting a fixed key sequence to injecting the modifier keys from the host sequence when we remove focus from a virtual machine and a key is pressed.  We do this before the key release event.  The idea is specifically so that switching away from the machine with alt-tab will not cause the guest to see a single alt key press and release, and activate a menu in some guest application.  We switched to using the host sequence because the user can choose it to be something which will not cause problems in the guest.
------------------------------------------------------------------------