commit_62397 ------------------------------------------------------------------------ r62397 | vboxsync | 2016-07-21 21:20:16 +0800 (Thu, 21 Jul 2016) | 2 lines
bugref:8087: Additions/Installer: On Debian-based systems we were trying to include vboxvideo in the initramfs even if we were uninstalling.  I do not know if this was serious, but check for it.  On dracut-based systems we were including the wrong (current) kernel modules and files in initrd files for non-current kernels.  Fix by passing the kernel version in the right place.

------------------------------------------------------------------------