commit_63064 ------------------------------------------------------------------------ r63064 | vboxsync | 2016-08-06 05:35:09 +0800 (Sat, 06 Aug 2016) | 1 line
vboxhgmc.c/_crVBoxHGSMIInit: VBoxCrHgsmiInit changed to a single signature without any parameters in r79662 & r79665, this code was forgotten because someone didn't know that one have to use put 'void' in empty parameter lists in C to form a real prototype.  Part of the blame should be given to using the VBOXR3NP template which doesn't make GCC warn about it.  Accidentally found when locally enabling this warning for MSC recently.
------------------------------------------------------------------------