commit_63156 ------------------------------------------------------------------------ r63156 | vboxsync | 2016-08-08 20:11:53 +0800 (Mon, 08 Aug 2016) | 1 line
GuestSession::i_objectCreateTempInternal,GuestSession::i_fsQueryInfoInternal: guestRc _was_ used uninitialized and pGuestRc not set when returning VERR_GSTCTL_GUEST_ERROR (local variable was updated instead). Changed procInfo initialization to return immediately when out of memory, as it makes the code flow unnecessarily complicated to do the single return point thing. The mExecutable was not inside the try/catch, as it need to be as it allocates memory too.
------------------------------------------------------------------------