commit_63491 ------------------------------------------------------------------------ r63491 | vboxsync | 2016-08-16 00:41:16 +0800 (Tue, 16 Aug 2016) | 9 lines
iprt/types.h: On NetBSD after we include <stdbool.h> undef "bool" it
defines and provide a typedef instead.  This is necessary because some
sloppy X11 code uses "bool" as a structure member name.

The end result is the same as for Darwin and Haiku clause just below,
but I think that explicit #include <stdbool.h> here (vs. checking for
its guard) is more clean, since it ensures that any code that includes
<stdbool.h> later will not cause additionall damage.

------------------------------------------------------------------------