commit_63558 ------------------------------------------------------------------------ r63558 | vboxsync | 2016-08-16 21:51:47 +0800 (Tue, 16 Aug 2016) | 7 lines
r0drv/netbsd: re-import memobj-r0drv-netbsd.c on top of an svn copy of
the FreeBSD version it's based on.  Since kernel vocabularies are
different, there's quite a bit of diff, but the general shape is more
or less the same to make it useful to keep the FreeBSD history.

From Haomai Wang GSoC project with additional changes by Arto Huusko.

------------------------------------------------------------------------