commit_63639 ------------------------------------------------------------------------ r63639 | vboxsync | 2016-08-25 22:31:10 +0800 (Thu, 25 Aug 2016) | 1 line
VBoxC: Looks like VirtualBoxClient::Data::m_pEventSource is holding in extra reference on the ATL module, causing VBoxC to be unloaded too late because DllCanUnloadNow() returns S_FALSE due to GetLockCount() being 1 instead of 0.  Created a hack for discounting the module lock held by the EventSource in the VirtualBoxClient singelton.
------------------------------------------------------------------------