commit_63772 ------------------------------------------------------------------------ r63772 | vboxsync | 2016-09-09 04:28:08 +0800 (Fri, 09 Sep 2016) | 1 line
bugref:8087: Additions/x11: support non-root X server: do not disable VBVA on mode changes, as this caused all our reported capability flags to be lost, and the host to think we did not support dynamic resizing.  We already disable VBVA when the master is changed and that is enough.  The motivation for disabling VBVA was that we did not enable it until we saw that the user-space master was willing to send us dirty rectangle information.
------------------------------------------------------------------------