commit_63779 ------------------------------------------------------------------------ r63779 | vboxsync | 2016-09-09 22:11:40 +0800 (Fri, 09 Sep 2016) | 1 line
bugref:8151: FE/Qt: improve X11 keyboard capturing.  We currently intercept mouse clicks outside of the machine window when the keyboard is grabbed so that we can release the keyboard grab in time to prevent it interfering with window drags and other operations.  Fixes: only grab the mouse button if the cursor is outside of the machine views to prevent interfering with input to the machine; and catch a few cases where the mouse button grab was not released.
------------------------------------------------------------------------