commit_63916 ------------------------------------------------------------------------ r63916 | vboxsync | 2016-09-21 04:26:22 +0800 (Wed, 21 Sep 2016) | 2 lines
bugref:8087: Additions/x11: support non-root X server: remove protection against loading while the X server is running which required that all virtual consoles be in text mode. We hopefully no longer need this, as the driver is only ever auto-loaded at boot time (particularly not when the Additions are installed), and it prevented the drivers from working in combination with EFI and a boot splash screen, as this also set the virtual console to graphics mode.

------------------------------------------------------------------------