commit_64181 ------------------------------------------------------------------------ r64181 | vboxsync | 2016-10-09 16:43:26 +0800 (Sun, 09 Oct 2016) | 1 line
bugref:8151: FE/Qt: improve X11 keyboard capturing: a user reported in ticketref:16046 that right-click context menus were broken in a number of panel tools when keyboard capturing was enabled.  Quick investigation showed that we were only grabbing the first mouse button for the purposes of doing a quick release of the keyboard capture when it was pressed.  This change makes us grab all mouse buttons instead.  I was able to verify locally that the user-described problem was reproducible before but not after the change.
------------------------------------------------------------------------